+++ plugins/modules/zhmc_crypto_attachment.py
@@ -82,14 +82,14 @@ options:
     required: true
   state:
     description:
-      - "The desired state for the attachment:"
+      - "The desired state for the crypto attachment. All states are fully
+         idempotent within the limits of the properties that can be changed:"
       - "* C(attached): Ensures that the specified number of crypto adapters
          of the specified crypto type, and the specified range of domain index
          numbers in the specified access mode are attached to the partition."
       - "* C(detached): Ensures that no crypto adapter and no crypto domains
          are attached to the partition."
-      - "* C(facts): Does not change anything on the attachment and returns
-         the crypto configuration of the partition."
+      - "* C(facts): Returns the crypto configuration of the partition."
     type: str
     required: true
     choices: ['attached', 'detached', 'facts']