+++ plugins/modules/zhmc_storage_group_attachment.py
@@ -92,14 +92,16 @@ options:
     required: true
   state:
     description:
-      - "The desired state for the attachment:"
+      - "The desired state for the storage group attachment. All states are
+         fully idempotent within the limits of the properties that can be
+         changed, unless otherwise stated:"
       - "* C(detached): Ensures that the storage group is not attached to the
          partition. If the storage group is currently attached to the partition
-         and the partition is currently active, the module will fail."
+         and the partition is currently active, the module will fail (this is
+         an idempotency limitation)."
       - "* C(attached): Ensures that the storage group is attached to the
          partition."
-      - "* C(facts): Does not change anything on the attachment and returns
-         the attachment status."
+      - "* C(facts): Returns the attachment status."
     type: str
     required: true
     choices: ['detached', 'attached', 'facts']